Trait tidb_query_executors::interface::ExecSummaryCollector

pub trait ExecSummaryCollector: Send {
    type DurationRecorder;
    pub fn new(output_index: usize) -> Self;

    pub fn on_start_iterate(&mut self) -> Self::DurationRecorder;

    pub fn on_finish_iterate(&mut self, dr: Self::DurationRecorder, rows: usize);

    pub fn collect(&mut self, target: &mut [ExecSummary]);
}

A trait for all execution summary collectors.

Associated Types

type DurationRecorder

Loading content...
Required methods

pub fn new(output_index: usize) -> Self

Creates a new instance with specified output slot index.

pub fn on_start_iterate(&mut self) -> Self::DurationRecorder

Returns an instance that will record elapsed duration and increase the iterations counter. The instance should be later passed back to on_finish_iterate when processing of next_batch is completed.

pub fn on_finish_iterate(&mut self, dr: Self::DurationRecorder, rows: usize)

pub fn collect(&mut self, target: &mut [ExecSummary])

Takes and appends current execution summary into target.

Loading content...